+++ tests/video.rs
@@ -55,6 +55,12 @@ fn test_mpeg4_part14_video() {
 }
 
 #[test]
+fn test_mpeg2_transport_stream() {
+    let format = FileFormat::from_file("fixtures/video/sample.mts").unwrap();
+    assert_eq!(format, FileFormat::Mpeg2TransportStream);
+}
+
+#[test]
 fn test_ogg_media() {
     let format = FileFormat::from_file("fixtures/video/sample.ogm").unwrap();
     assert_eq!(format, FileFormat::OggMedia);